requirements.txt
Flask==0.12.2
Pillow==5.2.0
requests==2.19.1
google-cloud-language==1.0.2